move construct_adt_metadata into GlobalContext
